Prometheus/Prometheus_example/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
fastapi==0.112.0
h11==0.14.0
idna==3.7
prometheus_client==0.20.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.37.2
typing==3.7.4.3
typing_extensions==4.12.2
uvicorn==0.30.6
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp

